+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Changed
 - CPX-AP: Added diagnosis example
+### Added
+- boollist_to_int convenience function
+- CLI: cpx_ap read/write channel functionality
+### Fixed
+- ApModule: write_channel bug causing other channels to be resetted for some module types. 
 
 ## v0.5.0 - 17.06.24
 ### Added